requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.17.14
botocore==1.20.14
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
MarkupSafe==2.0.1
packaging==21.3
psycopg2==2.9.2
psycopg2-binary==2.9.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.3
regex==2021.11.10
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.3.7
sqlparse==0.4.2
threadpoolctl==3.0.0
tqdm==4.62.3
uritemplate==4.1.1
urllib3==1.26.7
Pillow==8.4.0
django-storages==1.12.3
Werkzeug==2.0.2
